Adds PlaceIDCoordsRouter for coord lookup
